Initial Commit
This is a containerized copy of Smarden Runit.  The primary intent is
to be able to use the statically linked tools from runit to obviate
the need to restart the containing container, which doesn't always
work in certain environments.

Patches retrieved from the Void Linux project which is licensed
BSD-2-Clause.